# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Laravel Api",
        "version": "10.0"
    },
    "paths": {
...
        "/api/attendance-periodic-report/get-employee": {
            "post": {
                "tags": [
                    "Attendance Report"
                ],
                "summary": "GET The Employee.Permission required = attendance-report.list",
                "description": "This endpoint gives a specific employee. You just need to enter the emirates id of the employee, and it will return you the employee.",
                "operationId": "d2cd88c0df4e3c0736fbf8646a4c1249",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "emirates_id": {
                                        "description": "The emirates_id of the employee (required)",
                                        "type": "number",
                                        "example": "12121211212"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/attendance-periodic-report": {
            "post": {
                "tags": [
                    "Attendance Report"
                ],
                "summary": "GET attendance-periodic-report.Permission required = attendance-report.list",
                "description": "This endpoint get  attendance-periodic-report.",
                "operationId": "92b014be2be3d03717530957fa57fca0",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "from_date": {
                                        "description": "The date from which you want to start the report => required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-11"
                                    },
                                    "to_date": {
                                        "description": "The date to which you want to end the report => required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-11"
                                    },
                                    "employee_id": {
                                        "description": "the id of the employee for which you want to get the report",
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "attendance-daily-report created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/attendance-daily-report": {
            "post": {
                "tags": [
                    "Attendance Report"
                ],
                "summary": "GET attendance-daily-report.Permission required = attendance-report.list",
                "description": "This endpoint gets attendance-daily-report.",
                "operationId": "378f612fe8e9325de9c8cb687cdda8ae",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "description": "The date of the attendance-daily-report => required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-11"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "attendance-daily-report created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}